refactor(warehouse): one world-space convention for the gantry's traverse height
clearanceYFor mixed four conventions in a single expression: a rack
*surface* (`rackY + 1`), a rounded-up world position, and two raw block
indices - then handed the result to GantryState.moveTo as-is, while
every other waypoint moveTo builds is a Vec3.atCenterOf block centre.

Two things fell out of that. The traverse ran half a block below the
rail it was nominally following, and the head rode with its centre level
with the tallest rack's top surface - so its lower half (it's a 10x10x10
model centred on its position) sat inside the very rack it was meant to
be clearing.

Everything is now a world-space Y for the head's centre, which is what
moveTo actually consumes - clearanceY becomes a Double, and HEAD_CLEARANCE
names the half-head-plus-margin that keeps it above a rack rather than
level with it. The ceiling is expressed as the rail's own centre minus
that same clearance, so the head still hangs below the beams instead of
inside them - which works out to exactly the raw `bounds.max.y` this used
before, now by construction rather than by accident.

136/136 gametests pass, including the tall-bounds put-away case that
depends on this not ascending further than the racks require.

+++ common/src/main/kotlin/net/kernelpanicsoft/boilerplate/warehouse/GantryState.kt
@@ -12,7 +12,9 @@ import net.minecraft.world.phys.Vec3
  * Motion is deliberately not general 3D pathfinding: [moveTo] always resolves to a single ascent
  * to [clearanceY] (if not already there), then horizontal motion along X, then along Z, then a
  * final vertical descent/ascent onto the target - an idealized industrial gantry confined to its
- * rail envelope, not a voxel path through the player's build. [clearanceY] is the caller's own
+ * rail envelope, not a voxel path through the player's build. [clearanceY] is a world-space Y for
+ * the head's own *centre*, the same space [pos] and every waypoint are in - not a block index. It is
+ * the caller's own
  * responsibility to pick high enough to clear whatever's actually between the two points -
  * [WarehouseControllerBlockEntity] picks the lowest height that clears every rack it currently
  * knows about rather than always the bound volume's own top, since those can differ enormously (a
@@ -42,14 +44,14 @@ class GantryState(startPos: Vec3) {
 	 * back didn't, so the head climbed up and stayed there - see the early return below, which now
 	 * makes that case move nothing at all rather than bob up and back.
 	 */
-	fun moveTo(target: BlockPos, clearanceY: Int) {
+	fun moveTo(target: BlockPos, clearanceY: Double) {
 		val destination = Vec3.atCenterOf(target)
 		if (destination == pos) {
 			waypoints = ArrayDeque()
 			return
 		}
 
-		val rail = Vec3(pos.x, clearanceY.toDouble(), pos.z)
+		val rail = Vec3(pos.x, clearanceY, pos.z)
 		val overDestination = Vec3(destination.x, rail.y, destination.z)
 		val path = listOf(
 			rail,
